requirements.txt
google-adk
psutil
litellm
google-generativeai
python-dotenv
